+++ src/app/(home)/page.tsx
@@ -1,10 +1,11 @@
 import type { Metadata } from 'next'
 
 import { Divider } from '@/components/common'
-import HighlightedProjects from '@/components/highlighted-projects'
-import Introductory from '@/components/introductory'
-import RecentPosts from '@/components/recent-posts'
 import { getMetadata } from '@/lib/metadata'
+
+import HighlightedProjects from './highlighted-projects'
+import Introductory from './introductory'
+import RecentPosts from './recent-posts'
 
 export const metadata: Metadata = getMetadata()
